separate plastiticty and elasticity
